+++ warn/BUILD.bazel
@@ -5,10 +5,11 @@ go_library(
     srcs = [
         "types.go",
         "warn.go",
-        "warn_api.go",
-        "warn_appearance.go",
-        "warn_execution.go",
-        "warn_misc.go",
+        "warn_bazel.go",
+        "warn_bazel_misc.go",
+        "warn_bazel_syntax.go",
+        "warn_control_flow.go",
+        "warn_cosmetic.go",
         "warn_operation.go",
         "warn_syntax.go",
     ],
@@ -27,10 +28,11 @@ go_test(
     size = "small",
     srcs = [
         "types_test.go",
-        "warn_api_test.go",
-        "warn_appearance_test.go",
-        "warn_execution_test.go",
-        "warn_misc_test.go",
+        "warn_bazel_misc_test.go",
+        "warn_bazel_syntax_test.go",
+        "warn_bazel_test.go",
+        "warn_control_flow_test.go",
+        "warn_cosmetic_test.go",
         "warn_operation_test.go",
         "warn_syntax_test.go",
         "warn_test.go",
