+++ warn/warn_bazel.go
@@ -2,7 +2,6 @@ package warn
 
 import (
 	"fmt"
-
 	"github.com/bazelbuild/buildtools/build"
 	"github.com/bazelbuild/buildtools/bzlenv"
 	"github.com/bazelbuild/buildtools/edit"
